refactor : place JSONTrack interface in commons folder to share the same file between front and back
