+++ commons/Backend-types.ts
@@ -1,0 +1,8 @@
+export interface JSONTrack {
+  url: string;
+  title: string;
+  duration: number;
+  artistName: string;
+  albumName: string;
+  imgUrl: string;
+}
